+++ hugo/content/post/history.md
@@ -1,6 +1,6 @@
 ---
 title: history
-date: 2019-03-12
+date: 2019-03-13
 ---
 
 ## history
@@ -8,6 +8,7 @@ date: 2019-03-12
 ### -
 
 * [CLI Reference](https://docs.aws.amazon.com/cli/latest/reference/history/index.html)
+* Cmdlet Reference
 
 |AWS CLI|PowerShell Cmdlet|Prefix|
 |----|----|:--:|